+++ README.md
@@ -17,11 +17,11 @@ To install, run the following command:
 go get github.com/josephspurrier/goversioninfo/cmd/goversioninfo
 ~~~
 
-Copy versioninfo.json into your working directory and then modify the file with your own settings.
+Copy testdata/resource/versioninfo.json into your working directory and then modify the file with your own settings.
 
 Add a similar text to the top of your Go source code (-icon and -manifest are optional, but can also be specified in the versioninfo.json file):
 ~~~ go
-//go:generate goversioninfo -icon=icon.ico -manifest=goversioninfo.exe.manifest
+//go:generate goversioninfo -icon=testdata/resource/icon.ico -manifest=testdata/resource/goversioninfo.exe.manifest
 ~~~
 
 Run the Go commands in this order so goversioninfo will create a file called resource.syso in the same directory as the Go source code.
@@ -94,7 +94,7 @@ If you find any other differences, let me know.
 
 You can also use [windres](https://sourceware.org/binutils/docs/binutils/windres.html) to create the syso file. The windres executable is available in either [MinGW](http://www.mingw.org/) or [tdm-gcc](http://tdm-gcc.tdragon.net/).
 
-Below is a sample batch file you can use to create a .syso file from a .rc file. There are sample .rc files in the rc folder.
+Below is a sample batch file you can use to create a .syso file from a .rc file. There are sample .rc files in the testdata/rc folder.
 
 ~~~
 @ECHO OFF
