+++ testdata/example1/goversioninfo.exe.manifest
