+++ testdata/example1/main.go
@@ -1,0 +1,9 @@
+//go:generate goversioninfo -icon=icon.ico -manifest=goversioninfo.exe.manifest
+
+package main
+
+import "fmt"
+
+func main() {
+	fmt.Println("Hello world")
+}
