+++ testdata/example1/versioninfo.json
@@ -1,0 +1,43 @@
+{
+    "FixedFileInfo": {
+        "FileVersion": {
+            "Major": 1,
+            "Minor": 0,
+            "Patch": 0,
+            "Build": 0
+        },
+        "ProductVersion": {
+            "Major": 1,
+            "Minor": 0,
+            "Patch": 0,
+            "Build": 0
+        },
+        "FileFlagsMask": "3f",
+        "FileFlags ": "00",
+        "FileOS": "040004",
+        "FileType": "01",
+        "FileSubType": "00"
+    },
+    "StringFileInfo": {
+        "Comments": "Test file.",
+        "CompanyName": "josephspurrier",
+        "FileDescription": "This is a hello world file.",
+        "FileVersion": "v1.0.0.0",
+        "InternalName": "goversioninfo.exe",
+        "LegalCopyright": "Copyright (c) 2019 Joseph Spurrier",
+        "LegalTrademarks": "",
+        "OriginalFilename": "main.go",
+        "PrivateBuild": "",
+        "ProductName": "goversioninfo",
+        "ProductVersion": "v1.0.0.0",
+        "SpecialBuild": ""
+    },
+    "VarFileInfo": {
+        "Translation": {
+            "LangID": "0409",
+            "CharsetID": "04B0"
+        }
+    },
+    "IconPath": "",
+    "ManifestPath": ""
+}
