+++ testdata/rc/cmd.rc
