+++ testdata/rc/control.rc
