+++ testdata/rc/explorer.rc
