+++ testdata/rc/versioninfo.rc
