+++ testdata/resource/icon.ico
