+++ testdata/resource/versioninfo.json
